db: Abstract sqlite into generic sql
